adk/requirements.txt
# IntentVision ADK Requirements
# Following bobs-brain Hard Mode (R1: ADK-only)
# Beads Task: intentvision-qd3

# Core ADK (R1: Agent implementation framework)
google-adk>=1.18.0,<1.19.0

# A2A Protocol (R7: Agent-to-Agent communication)
a2a-sdk>=0.3.0

# Google Cloud Platform
google-cloud-aiplatform>=1.112.0
google-auth>=2.23.0
google-cloud-logging>=3.8.0

# Gateway dependencies (service/)
fastapi>=0.104.0
httpx>=0.25.0
pydantic>=2.4.0
uvicorn[standard]>=0.24.0

# Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0

# Utilities
python-dotenv>=1.0.0
